requirements.txt
pygame-ce
numpy
scipy
numba